Initial add
Basic use of template functions, go1.6 template blocks.

Not sure of the best approach, a global single massive *template.Template with
all of the named templates within... or what I've gone for which is each actual
page to be displayed is a distinct template tree within a
map[string]*template.Template.

I prefer my approach as I generally go for declarativeness as early as possible
and it being abundantly clear later which template files are actually used to
display which pages. As blocks get more heavily used this would seem wiser.

Future work should look into how to have structs have their own render functions
and how to pass the output of that into a template. Perhaps these are
pre-registered funcs added to the funcMap that takes in a strongly typed struct
and returns the HTML... which I then need to mark safe?

The scenario for that would be search results. Results may include lots of types
of things, and rather than having massive switch, I'd rather say
 "thing, render thyself" and for that to just work.

Anyhow... basics of funcmap + blocks + nesting all working with basic HTML
minimisation included.